Fix schema mapping for duplicate fields.
We used to map each schema element from their names to the element type. This is wrong, because for nested fields, hashmap was using a duplicate key, resulting in overwriting the previous entry. For instance, f1.list.element and f2.list.element would be mapped only once, using the name 'element'. This commit fixes it.